Day 4: Spring Boot wired end-to-end — security, exception handling, Testcontainers integration tests
- SecurityConfig: stateless, CSRF off, all requests permitted (JWT added Day 7)
  BCryptPasswordEncoder bean registered
- Exception hierarchy: ResourceNotFoundException (404), BusinessRuleException (422)
- ErrorResponse record with status/code/message/timestamp envelope
- GlobalExceptionHandler: @RestControllerAdvice maps all exception types
- Testcontainers: PostgreSQLContainer + @ServiceConnection, shared across test classes
- FlywayMigrationTest: 28 assertions — tables, columns, constraints, indexes, CHECKs
- JavaApiApplicationTests: context-load smoke test
- pom.xml: micrometer-prometheus, testcontainers deps, colima auto-profile
  (DOCKER_HOST, TESTCONTAINERS_RYUK_DISABLED, api.version=1.44)
- V1__init.sql: seat_no + dealer_seat SMALLINT → INT (Hibernate type alignment)
- ~/.testcontainers.properties: tc.host + ryuk.disabled for Colima socket

+++ java-api/src/main/resources/db/migration/V1__init.sql
@@ -45,7 +45,7 @@ CREATE TABLE table_seats (
     id          UUID      NOT NULL DEFAULT gen_random_uuid(),
     table_id    UUID      NOT NULL,
     player_id   UUID,                        -- NULL = empty seat
-    seat_no     SMALLINT  NOT NULL,
+    seat_no     INT       NOT NULL,
     stack_chips INT       NOT NULL DEFAULT 0,
     sitting_out BOOLEAN   NOT NULL DEFAULT FALSE,
 
@@ -61,7 +61,7 @@ CREATE TABLE table_seats (
 CREATE TABLE hands (
     id          UUID        NOT NULL DEFAULT gen_random_uuid(),
     table_id    UUID        NOT NULL,
-    dealer_seat SMALLINT    NOT NULL,
+    dealer_seat INT         NOT NULL,
     street      VARCHAR(20) NOT NULL DEFAULT 'PREFLOP',
     status      VARCHAR(20) NOT NULL DEFAULT 'WAITING',
     pot_chips   INT         NOT NULL DEFAULT 0,
